seperate testfiles for unit tests for ext. preds.
