+++ doubleml/tests/test_plr_external_predictions.py
@@ -2,17 +2,13 @@ import numpy as np
 import pytest
 import math
 from sklearn.linear_model import LinearRegression, LassoCV, LogisticRegression
-from doubleml import DoubleMLPLR, DoubleMLIRM, DoubleMLData
-from doubleml.datasets import make_plr_CCDDHNR2018, make_irm_data
-from doubleml.utils import dummy_regressor, dummy_classifier
+from doubleml import DoubleMLPLR, DoubleMLData
+from doubleml.datasets import make_plr_CCDDHNR2018
+from doubleml.utils import dummy_regressor
 
 
 @pytest.fixture(scope="module", params=["IV-type", "partialling out"])
 def plr_score(request):
-    return request.param
-
-@pytest.fixture(scope="module", params=["ATE", "ATTE"])
-def irm_score(request):
     return request.param
 
 
@@ -63,45 +59,6 @@ def doubleml_plr_fixture(plr_score, dml_procedure, n_rep):
     return res_dict
 
 
-@pytest.fixture(scope="module")
-def doubleml_irm_fixture(irm_score, dml_procedure, n_rep):
-    ext_predictions = {"d": {}}
-
-    x, y, d = make_irm_data(n_obs=500, dim_x=20, theta=0.5, return_type="np.array")
-
-    np.random.seed(3141)
-
-    dml_data = DoubleMLData.from_arrays(x=x, y=y, d=d)
-
-    kwargs = {"obj_dml_data": dml_data, "score": irm_score, "n_rep": n_rep, "dml_procedure": dml_procedure}
-
-    DMLIRM = DoubleMLIRM(ml_g=LinearRegression(), ml_m=LogisticRegression(), **kwargs)
-    np.random.seed(3141)
-
-    DMLIRM.fit(store_predictions=True)
-
-    ext_predictions["d"]["ml_g0"] = DMLIRM.predictions["ml_g0"][:, :, 0]
-    ext_predictions["d"]["ml_g1"] = DMLIRM.predictions["ml_g1"][:, :, 0]
-    ext_predictions["d"]["ml_m"] = DMLIRM.predictions["ml_m"][:, :, 0]
-
-    DMLIRM_ext = DoubleMLIRM(ml_g=dummy_regressor(), ml_m=dummy_classifier(), **kwargs)
-
-    np.random.seed(3141)
-    DMLIRM_ext.fit(external_predictions=ext_predictions)
-
-    res_dict = {"coef_normal": DMLIRM.coef, "coef_ext": DMLIRM_ext.coef}
-
-    return res_dict
-
-
 @pytest.mark.ci
 def test_doubleml_plr_coef(doubleml_plr_fixture):
-    assert math.isclose(
-        doubleml_plr_fixture["coef_normal"], doubleml_plr_fixture["coef_ext"], rel_tol=1e-9, abs_tol=1e-4
-    )
-    
-@pytest.mark.ci
-def test_doubleml_irm_coef(doubleml_irm_fixture):
-    assert math.isclose(
-        doubleml_irm_fixture["coef_normal"], doubleml_irm_fixture["coef_ext"], rel_tol=1e-9, abs_tol=1e-4
-    )
+    assert math.isclose(doubleml_plr_fixture["coef_normal"], doubleml_plr_fixture["coef_ext"], rel_tol=1e-9, abs_tol=1e-4)
